add commit log http server

+++ internal/server/log.go
@@ -25,7 +25,7 @@ func (c *Log) Append(record Record) (uint64, error) {
 func (c *Log) Read(offset uint64) (Record, error) {
 	c.mu.Lock()
 	defer c.mu.Unlock()
-	if offset >= len(c.records) {
+	if offset >= uint64(len(c.records)) {
 		return Record{}, ErrOffsetNotFound
 	}
 	return c.records[offset], nil
